Add delete test
