+++ test/CMakeLists.txt
@@ -25,6 +25,7 @@ include_directories(../rpcserver/include)
 include_directories(../storage/include)
 include_directories(util/include)
 include_directories(../web-request/include)
+include_directories(../web-rdfs/include)
 
 link_directories(/home/vagrant/rdfs/build/zkwrapper)
 
@@ -100,11 +101,10 @@ target_link_libraries(UsernameTest namenode_protos)
 target_link_libraries(UsernameTest ${GTEST})
 target_link_libraries(UsernameTest zk_client_nn rpcserver)
 
-add_executable(WebRequestTranslatorTest webRDFS/WebRequestTranslatorTest.cc)
-target_link_libraries(WebRequestTranslatorTest ${GTEST_LIBRARIES} pthread)
-target_link_libraries(WebRequestTranslatorTest ${GTEST})
-target_link_libraries(WebRequestTranslatorTest namenode_protos)
-target_link_libraries(WebRequestTranslatorTest zk_client_nn zk_client_dn rpcserver)
-target_link_libraries(WebRequestTranslatorTest zkwrapper /usr/local/lib/libzookeeper_mt.so)
-target_link_libraries(WebRequestTranslatorTest web-request)
-
+add_executable(WebServerTest webRDFS/WebServerTest.cc)
+target_link_libraries(WebServerTest ${GTEST_LIBRARIES} pthread)
+target_link_libraries(WebServerTest ${GTEST})
+target_link_libraries(WebServerTest namenode_protos)
+target_link_libraries(WebServerTest zk_client_nn zk_client_dn rpcserver)
+target_link_libraries(WebServerTest zkwrapper /usr/local/lib/libzookeeper_mt.so)
+target_link_libraries(WebServerTest web-request web_rdfs_server)
